make helper inner class
