+++ rls/src/test/java/io/grpc/rls/internal/LbPolicyConfigurationTest.java
@@ -16,23 +16,35 @@
 
 package io.grpc.rls.internal;
 
+import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import io.grpc.Attributes;
+import io.grpc.ConnectivityState;
+import io.grpc.ConnectivityStateInfo;
+import io.grpc.EquivalentAddressGroup;
+import io.grpc.LoadBalancer.CreateSubchannelArgs;
 import io.grpc.LoadBalancer.Helper;
+import io.grpc.LoadBalancer.Subchannel;
 import io.grpc.LoadBalancer.SubchannelPicker;
+import io.grpc.LoadBalancer.SubchannelStateListener;
 import io.grpc.LoadBalancerProvider;
 import io.grpc.LoadBalancerRegistry;
 import io.grpc.rls.internal.ChildLoadBalancerHelper.ChildLoadBalancerHelperProvider;
-import io.grpc.rls.internal.ChildPolicyReportingHelper.ChildLbStatusListener;
 import io.grpc.rls.internal.LbPolicyConfiguration.ChildLoadBalancingPolicy;
 import io.grpc.rls.internal.LbPolicyConfiguration.ChildPolicyWrapper;
+import io.grpc.rls.internal.LbPolicyConfiguration.ChildPolicyWrapper.ChildLbStatusListener;
+import io.grpc.rls.internal.LbPolicyConfiguration.ChildPolicyWrapper.ChildPolicyReportingHelper;
 import io.grpc.rls.internal.LbPolicyConfiguration.InvalidChildPolicyConfigException;
 import io.grpc.rls.internal.LbPolicyConfiguration.RefCountedChildPolicyWrapperFactory;
+import java.net.SocketAddress;
 import java.util.Map;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -41,13 +53,14 @@ import org.junit.runners.JUnit4;
 @RunWith(JUnit4.class)
 public class LbPolicyConfigurationTest {
 
+  private final Helper helper = mock(Helper.class);
+  private final SubchannelStateManager subchannelStateManager = new SubchannelStateManagerImpl();
+  private final SubchannelPicker picker = mock(SubchannelPicker.class);
+  private final ChildLbStatusListener childLbStatusListener = mock(ChildLbStatusListener.class);
   private final RefCountedChildPolicyWrapperFactory factory =
       new RefCountedChildPolicyWrapperFactory(
-          new ChildLoadBalancerHelperProvider(
-              mock(Helper.class),
-              mock(SubchannelStateManager.class),
-              mock(SubchannelPicker.class)),
-          mock(ChildLbStatusListener.class));
+          new ChildLoadBalancerHelperProvider(helper, subchannelStateManager, picker),
+          childLbStatusListener);
 
   @Test
   public void childPolicyWrapper_refCounted() {
@@ -117,6 +130,71 @@ public class LbPolicyConfigurationTest {
       fail("parsing exception expected");
     } catch (InvalidChildPolicyConfigException e) {
       assertThat(e).hasMessageThat().contains("no valid childPolicy found");
+    }
+  }
+
+  @Test
+  public void subchannelStateChange_updateChildPolicyWrapper() {
+    ChildPolicyWrapper childPolicyWrapper = factory.createOrGet("foo.google.com");
+    ChildPolicyReportingHelper childPolicyReportingHelper = childPolicyWrapper.getHelper();
+    FakeSubchannel fakeSubchannel = new FakeSubchannel();
+    when(helper.createSubchannel(any(CreateSubchannelArgs.class))).thenReturn(fakeSubchannel);
+    Subchannel subchannel =
+        childPolicyReportingHelper
+            .createSubchannel(
+                CreateSubchannelArgs.newBuilder()
+                    .setAddresses(new EquivalentAddressGroup(mock(SocketAddress.class)))
+                    .build());
+    subchannel.start(new SubchannelStateListener() {
+      @Override
+      public void onSubchannelState(ConnectivityStateInfo newState) {
+        // no-op
+      }
+    });
+
+    fakeSubchannel.updateState(ConnectivityStateInfo.forNonError(ConnectivityState.CONNECTING));
+
+    assertThat(childPolicyWrapper.getConnectivityStateInfo())
+        .isEqualTo(ConnectivityStateInfo.forNonError(ConnectivityState.CONNECTING));
+  }
+
+  @Test
+  public void updateBalancingState_triggersListener() {
+    ChildPolicyWrapper childPolicyWrapper = factory.createOrGet("foo.google.com");
+    ChildPolicyReportingHelper childPolicyReportingHelper = childPolicyWrapper.getHelper();
+    SubchannelPicker childPicker = mock(SubchannelPicker.class);
+
+    childPolicyReportingHelper.updateBalancingState(ConnectivityState.READY, childPicker);
+
+    verify(childLbStatusListener).onStatusChanged(ConnectivityState.READY);
+    assertThat(childPolicyWrapper.getPicker()).isEqualTo(childPicker);
+    // picker governs childPickers will be reported to parent LB
+    verify(helper).updateBalancingState(ConnectivityState.READY, picker);
+  }
+
+  private static class FakeSubchannel extends Subchannel {
+
+    private SubchannelStateListener listener;
+
+    @Override
+    public void start(SubchannelStateListener listener) {
+      this.listener = listener;
+    }
+
+    void updateState(ConnectivityStateInfo newState) {
+      checkState(listener != null, "channel is not started yet");
+      listener.onSubchannelState(newState);
+    }
+
+    @Override
+    public void shutdown() {}
+
+    @Override
+    public void requestConnection() {}
+
+    @Override
+    public Attributes getAttributes() {
+      return null;
     }
   }
 }
